+++ src/commands.ts
@@ -8,6 +8,7 @@ import { CompileOption } from './languageServer/options';
 import { configuration } from './common/configuration'
 import { lazproject } from './common/lazproject'
 import { type } from 'os';
+import { EditorCommandManager } from './editor';
 import { client } from './extension';
 import { TextEditor, TextEditorEdit } from 'vscode';
 
@@ -36,7 +37,10 @@ export class FpcCommandManager {
 
         // General commands
         context.subscriptions.push(vscode.commands.registerTextEditorCommand('fpctoolkit.code.rename',this.CodeRename));
-        context.subscriptions.push(vscode.commands.registerTextEditorCommand('fpctoolkit.editor.trimfromcursor',this.TrimFromCursor));
+
+        // Register commands
+        let commands = new EditorCommandManager();
+        commands.registerAll(context);
     }
     ProjectAdd = async (node: FpcItem) => {
         if (node.level === 0) {
@@ -180,99 +184,6 @@ export class FpcCommandManager {
             return;
 
         lazproject.ProjectActivate(this.workspaceRoot, file);
-    };
-    TrimFromCursor = async (textEditor: TextEditor, edit: TextEditorEdit) => {
-
-        const editor = vscode.window.activeTextEditor;
-
-        if (editor) {
-            const cursorPosition = editor.selection.active; // Get current cursor position
-            const documentText = editor.document.getText();
-            const textAfterCursor = documentText.slice(editor.document.offsetAt(cursorPosition));
-            const textBeforeCursor = documentText.slice(editor.document.offsetAt(cursorPosition) - 1, editor.document.offsetAt(cursorPosition));
-
-            let deleteLength = 0;
-
-            // Helper function to determine if a character is a word character
-            const isWordChar = (char: string) => /[a-zA-Z0-9_#]/.test(char);
-
-            // Helper function to determine if a character is whitespace
-            const isWhitespace = (char: string) => /\s/.test(char);
-
-            // Helper function to determine if a character is a special character
-            const isSpecialChar = (char: string) => /[^\w\s]/.test(char);
-
-            // Helper function for CRLF detection
-            const isCRLF = (char: string) => char === '\n' || char === '\r';
-
-            // Initialize flags
-            let isStartWord = false;
-            let isBeforeWord = false;
-            let isStartCRLF = false;
-            if (textBeforeCursor.length>0)
-                isBeforeWord = isWordChar(textBeforeCursor[0]);
-            if (textAfterCursor.length>0) {
-                isStartWord = isWordChar(textAfterCursor[0]);
-                isStartCRLF = isCRLF(textAfterCursor[0]);
-            }
-
-            for (let i = 0; i < textAfterCursor.length; i++) {
-
-                const char = textAfterCursor[i];
-                const nextChar = textAfterCursor[i + 1] || '';
-
-                if (isWhitespace(char)) {
-                    // If in whitespace, delete spaces and tabs until the next non-whitespace character
-                    if (isCRLF(char)) {
-
-                        if (isStartCRLF && isWhitespace(nextChar) && !isCRLF(nextChar)) {
-                            isStartCRLF = false;
-                            continue;
-                        }
-
-                        // If a blank line is encountered, delete only the first one and stop
-                        deleteLength = i + 1;
-                        break;
-                    } else if (!isWhitespace(nextChar)) {
-                        deleteLength = i + 1;
-                        break;
-                    }
-                } else if (isWordChar(char)) {
-                    // If in a word, delete until the end of the word or a special character
-                    if (!isWordChar(nextChar)) {
-
-                        // The char before was not a word so we want to delete all the way to next whitespaces
-                        if (!isBeforeWord && isWhitespace(nextChar) && !isCRLF(nextChar))
-                            continue;
-
-                        deleteLength = i + 1;
-                        break;
-                    }
-                } else if (isSpecialChar(char)) {
-                    // If in special characters, delete until the group of special characters ends
-                    if (!isSpecialChar(nextChar)) {
-
-                        // Delete to all whitespaces
-                        if (isWhitespace(nextChar) && !isCRLF(nextChar))
-                            continue;
-
-                        deleteLength = i + 1;
-                        break;
-                    }
-                }
-            }
-
-            // Get the range to delete
-            const range = new vscode.Range(
-                cursorPosition,
-                editor.document.positionAt(editor.document.offsetAt(cursorPosition) + deleteLength)
-            );
-
-            // Perform the text replacement (deletion)
-            await editor.edit(editBuilder => {
-                editBuilder.delete(range);
-            });
-        }
     };
     ProjectNew = async () => {